add-payments-flow

+++ src/Core/Deliver.Application/Features/Trips/DriverTrips/Query/GetDriverCurrentTrip/DriverCurrentTripVm.cs
@@ -12,5 +12,5 @@ public class DriverCurrentTripVm
     public required DateTime CreatedDate { get; init; }
     public required double CalculatedDistance { get; init; }
     public required double CalculatedDuration { get; init; }
-    public double? CaptainProfit { get; set; }
+    public decimal? CaptainProfit { get; set; }
 }
